The PPD conformance tests did not catch group translation strings that exceeded the maximum allowed size (STR #454)
Updated the client code in the scheduler to close the client
connection on errors rather than shutting down the receive end
of the socket; this caused resource problems on some systems
(STR #434)

cups-polld didn't compile on Tru64 5.1B (STR #436)

"lpc stat" crashed if the device URI was empty (STR #548)

The scheduler did not compile without zlib (STR #433)

std:floor() cast needed on IRIX 6.5 with SGI C++ compiler (STR
#497)

cupsRasterReadPixels() and cupsRasterWritePixels() did not
handle EAGAIN and EINTR properly (STR #473)

RequiresPageRegion should not be consulted for Manual Feed (STR
#514)

International characters were not substituted in banner files
properly (STR #468)


git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@4033 7a7537e8-13f0-0310-91df-b6672ffda945

+++ pdftops/PSOutputDev.cxx
@@ -1933,8 +1933,8 @@ void PSOutputDev::startPage(int pageNum, GfxState *state) {
     x2 = (int)(state->getX2() + 0.5);
     y2 = (int)(state->getY2() + 0.5);
     writePSFmt("%%%%PageBoundingBox: %d %d %d %d\n",
-	       (int)floor(x1), (int)floor(y1),
-	       (int)ceil(x2), (int)ceil(y2));
+	       (int)floor((float)x1), (int)floor((float)y1),
+	       (int)ceil((float)x2), (int)ceil((float)y2));
     writePS("%%BeginPageSetup\n");
     width = x2 - x1;
     height = y2 - y1;
